Move CLI into cmd/client
